New SaveFileTwice test focusing on the Apache PDFBox Jira issue PDFBOX-5696
COSStream lost, becomes a COSDictionary
https://issues.apache.org/jira/browse/PDFBOX-5696
TemplateTank.pdf
https://issues.apache.org/jira/secure/attachment/13063398/TemplateTank.pdf
CambriaMath.ttf
https://issues.apache.org/jira/secure/attachment/13063399/CambriaMath.ttf

testSaveTwiceLikePadosAttila - Indeed, the appearance stream loses its stream data. Also important, it is included in an object stream. This might cause the loss of the stream data as streams are not allowed in object streams. Furthermore, another issue becomes apparent already in the first saved object: This single revision document has a cross reference stream with segmented object number ranges. This strictly speaking is not allowed in the first cross reference section of a PDF.