New CopyFormattedPageText tests focusing on the stack overflow question 77706995
How can I copy extracted text page by page in a document to a new PDF document with PDFBox?
https://stackoverflow.com/questions/77706995/how-can-i-copy-extracted-text-page-by-page-in-a-document-to-a-new-pdf-document-w

copyText - This method implements a way to copy text content from one document to another following the ideas presented by the OP.

Beware, this is just a proof of concept and not a complete implementation. In particular page rotation and non-upright text in general are not supported. Also the only supported style attributes are text font and text size, other details (like text color) are ignored. Different page geometries in source and target also will result in weird appearances.

The test methods testCopyFromElonInput, testCopyFromTemplateTank, and testCopyFromTestDocumentSigned apply the copyText method to a number of arbitrary documents.