Updated CopyFormattedPageText tests focusing on the stack overflow question 77706995
How can I copy extracted text page by page in a document to a new PDF document with PDFBox?
https://stackoverflow.com/questions/77706995/how-can-i-copy-extracted-text-page-by-page-in-a-document-to-a-new-pdf-document-w

copyText - updated this method to also accept a Consumer<List<TextPosition>> parameter to manipulate the extracted text position objects between retrieval and drawing.

searchAndReplace - This method replaces a search word in a list of TextPosition objects by replacing the letters in each instance by the same number of letters from the replacement word (as long as available). This is appropriate if the word is especially formatted (e.g. spaced out) and this special formatting shall be kept.

searchAndReplaceAlternative - This method replaces a search word in a list of TextPosition objects by replacing the letters in the first instance by the whole replacement word and removing the other instances. This is appropriate if the word is not formatted (e.g. spaced out) and shall be printed naturally.

The test methods testCopyAndChangeFromTestDocumentSigned and testCopyAndChangeAlternativelyFromTestDocumentSigned apply the copyText method with the above consumers to a document.